requirements.txt
django==1.11.22
Markdown==2.4
mysqlclient==1.4.2.post1
Pygments==1.6
bottle==0.12.9
cssmin==0.2.0
django-assets==0.12
django-debug-toolbar==1.8
django-celery==3.2.1
django-waffle==0.14.0
futures==2.1.6
gunicorn==0.17.4
django-hamlpy==1.1.1
httpretty==0.8.3
jsmin==2.2.1
mock==1.3.0
six==1.10.0
urllib3==1.15.1
webassets==0.12.1
libsass==0.13.3
boto>=2.1.0
django-ses==0.7.1
django-storages==1.6.6
Pillow==2.7.0
newrelic==4.14.0.115
python-memcached==1.59
pytz==2015.2
-e git+https://github.com/edx/edx-notifications.git@1.0.2#egg=edx-notifications==1.0.2
djangorestframework==3.6.3
django-upload-validator
requests_oauthlib==0.8.0
oauthlib>=2.1.0,<3
beautifulsoup4==4.6.0
flake8==3.6.0

# pylint plugins for django and celery
pylint-celery==0.3
pylint-django==0.7.2

# pylint dependencies
astroid==1.5.3
isort==4.2.15

pylint==1.7.2
user-agents==1.1.0

# Release utils for the edx release pipeline
edx-django-release-util==0.3.1

requests==2.18.4
unicodecsv==0.14.1

edx-i18n-tools==0.4.2

# Static i18n support
django-statici18n==1.8.3

# Redis version
redis==2.10.6
jsonfield==2.0.2
